+++ src/app/src/util/constants.js
@@ -53,6 +53,19 @@ export const facilityClaimStatusChoicesEnum = Object.freeze({
     REVOKED: 'REVOKED',
 });
 
+export const facilityListStatusChoicesEnum = Object.freeze({
+    PENDING: 'PENDING',
+    APPROVED: 'APPROVED',
+    REJECTED: 'REJECTED',
+    MATCHED: 'MATCHED',
+});
+export const facilityListStatusChoices = [
+    { value: facilityListStatusChoicesEnum.MATCHED, label: 'Matched' },
+    { value: facilityListStatusChoicesEnum.PENDING, label: 'Pending' },
+    { value: facilityListStatusChoicesEnum.APPROVED, label: 'Approved' },
+    { value: facilityListStatusChoicesEnum.REJECTED, label: 'Rejected' },
+];
+
 // These choices must be kept in sync with the identical list
 // kept in the Django API's constants.py file
 export const matchResponsibilityEnum = Object.freeze({
@@ -356,6 +369,7 @@ export const facilityListItemStatusChoicesEnum = Object.freeze({
     ERROR_MATCHING: 'ERROR_MATCHING',
     DELETED: 'DELETED',
     REMOVED: 'REMOVED', // This is not a status that appears in the database
+    DUPLICATE: 'DUPLICATE',
 });
 
 export const facilityListItemErrorStatuses = Object.freeze([